+++ src/app/api/organization/[id]/invite/route.ts
@@ -1,3 +1,4 @@
+import { headers } from 'next/headers';
 import { NextResponse } from 'next/server';
 
 import { API_BASE_URL } from '@/constant';
@@ -7,22 +8,23 @@ export async function POST(
   req: Request,
   { params }: { params: { id: string } }
 ) {
-  const {
-    token,
-    organizationName,
-    organizationUserEmail,
-    organizationAdminId,
-  } = await req.json();
+  const auth = headers().get('Authorization');
+  const { organizationName, organizationUserEmail, organizationAdminId } =
+    await req.json();
 
   const organizationId = params.id;
 
   try {
+    if (!auth) {
+      throw new Error(`Missing authorization header`);
+    }
+
     const response = await fetch(
       `${API_BASE_URL}/organization/${organizationId}/invite`,
       {
         method: 'POST',
         headers: {
-          Authorization: `Bearer ${token}`,
+          Authorization: auth,
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
@@ -57,18 +59,23 @@ export async function DELETE(
   req: Request,
   { params }: { params: { id: string } }
 ) {
-  const { token, organizationName, organizationUserId, organizationAdminId } =
+  const auth = headers().get('Authorization');
+  const { organizationName, organizationUserId, organizationAdminId } =
     await req.json();
 
-  const organizationId = params.id;
+  const userId = params.id;
 
   try {
+    if (!auth) {
+      throw new Error(`Missing authorization header`);
+    }
+
     const response = await fetch(
-      `${API_BASE_URL}/organization/${organizationId}/invite`,
+      `${API_BASE_URL}/organization/${userId}/invite`,
       {
         method: 'DELETE',
         headers: {
-          Authorization: `Bearer ${token}`,
+          Authorization: auth,
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
